handle all types in service deployment properties
